feat(): fix bug

+++ api-job.go
@@ -16,7 +16,7 @@ func (s *Server) broadcast(c *gin.Context) {
 		return
 	}
 	// todo whether can view has being processed?
-	if err := s.jobManager.RegisterJob(arid, jobTypeBroadcast, int64(len(s.peers))); err != nil {
+	if err := s.jobManager.RegisterJob(arid, jobTypeBroadcast, len(s.peers)); err != nil {
 		c.JSON(http.StatusBadGateway, err.Error())
 		return
 	}
@@ -45,7 +45,7 @@ func (s *Server) sync(c *gin.Context) {
 		return
 	}
 
-	if err := s.jobManager.RegisterJob(arid, jobTypeSync, int64(len(s.peers))); err != nil {
+	if err := s.jobManager.RegisterJob(arid, jobTypeSync, len(s.peers)); err != nil {
 		c.JSON(http.StatusBadGateway, err.Error())
 		return
 	}
